fix for vue
