+++ server/templates/login.html
@@ -6,6 +6,6 @@
 <title>Neotter Login</title>
 </head>
 <body>
-    <a href={{endpoint}}>Twitter連携する / Twitter integration</p>
+    <a href=((endpoint))>Twitter連携する / Twitter integration</p>
 </body>
 </html>
